requirements.txt
asgiref==3.8.1
boto3==1.35.44
botocore==1.35.44
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-rosetta==0.10.1
django-solo==2.3.0
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
jmespath==1.0.1
polib==1.2.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
